Fix example images being created with the wrong suffix
When generate_examples.py was migrated to use Python's pathlib in
4ee9761d, the suffixes of generated example images (in
assets/images/GeneratedImages) went from being ".mei.svg" to just ".svg"
because Path.with_suffix() replaces, rather than appends to, the
existing suffix. This caused references to those files on the site to be
broken.

This change alters the xslt generating the image link references to use
.svg as well, restoring those links.

Refs #1458